test(agent): split RFC-64 control store suites
