+++ packages/agent/vitest.unit.config.ts
@@ -116,6 +116,9 @@ export default defineConfig({
       "test/rfc64-agent-inventory-lifecycle.test.ts",
       "test/rfc64-author-catalog-producer.test.ts",
       "test/rfc64-control-object-store-v1.test.ts",
+      "test/rfc64-control-object-store-lifecycle-v1.test.ts",
+      "test/rfc64-durable-file-store-v1.test.ts",
+      "test/rfc64-secure-filesystem-policy-v1.test.ts",
     ],
     testTimeout: 60_000,
     maxWorkers: 1,
